[18] - Update function saveWalletInCache location
