+++ app/Application/DataSources/WalletDataSource.php
@@ -12,4 +12,6 @@ interface WalletDataSource
     public function insertCoinInWallet(string $wallet_id, Coin $coin): void;
 
     public function sellCoinFromWallet(string $wallet_id, Coin $coin, float $updatedUsdValue, string $amoungUsd): void;
+
+    public function saveWalletIncache(): ?string;
 }
